Fix Rubocop errors
